+++ src/roq/udp_subscriber/incremental.hpp
@@ -11,10 +11,11 @@
 
 #include "roq/io/net/udp/receiver.hpp"
 
-#include "roq/udp_subscriber/buffer.hpp"
 #include "roq/udp_subscriber/parser.hpp"
 #include "roq/udp_subscriber/reader.hpp"
 #include "roq/udp_subscriber/shared.hpp"
+
+#include "roq/udp_subscriber/tools/buffer.hpp"
 
 namespace roq {
 namespace udp_subscriber {
@@ -48,21 +49,21 @@ class Incremental final : public io::net::udp::Receiver::Handler, public Parser:
   void operator()(io::net::udp::Receiver::Error const &) override;
 
   // Parser::Handler
-  void operator()(Trace<Parser::Heartbeat> const &, Header const &) override;
-  void operator()(Trace<GatewaySettings> const &, Header const &) override;
-  void operator()(Trace<StreamStatus> const &, Header const &) override;
-  void operator()(Trace<ExternalLatency> const &, Header const &) override;
-  void operator()(Trace<GatewayStatus> const &, Header const &) override;
-  void operator()(Trace<ReferenceData> const &, Header const &) override;
-  void operator()(Trace<MarketStatus> const &, Header const &) override;
-  void operator()(Trace<TopOfBook> const &, Header const &) override;
-  void operator()(Trace<MarketByPriceUpdate> const &, Header const &) override;
-  void operator()(Trace<TradeSummary> const &, Header const &) override;
-  void operator()(Trace<StatisticsUpdate> const &, Header const &) override;
-  void operator()(Trace<CustomMetricsUpdate> const &, Header const &) override;
+  void operator()(Trace<Parser::Heartbeat> const &, tools::Header const &) override;
+  void operator()(Trace<GatewaySettings> const &, tools::Header const &) override;
+  void operator()(Trace<StreamStatus> const &, tools::Header const &) override;
+  void operator()(Trace<ExternalLatency> const &, tools::Header const &) override;
+  void operator()(Trace<GatewayStatus> const &, tools::Header const &) override;
+  void operator()(Trace<ReferenceData> const &, tools::Header const &) override;
+  void operator()(Trace<MarketStatus> const &, tools::Header const &) override;
+  void operator()(Trace<TopOfBook> const &, tools::Header const &) override;
+  void operator()(Trace<MarketByPriceUpdate> const &, tools::Header const &) override;
+  void operator()(Trace<TradeSummary> const &, tools::Header const &) override;
+  void operator()(Trace<StatisticsUpdate> const &, tools::Header const &) override;
+  void operator()(Trace<CustomMetricsUpdate> const &, tools::Header const &) override;
 
   template <typename T>
-  bool update(Trace<T> const &, Header const &);
+  bool update(Trace<T> const &, tools::Header const &);
 
   void publish_stream_status(TraceInfo const &, ConnectionStatus);
 
@@ -75,7 +76,7 @@ class Incremental final : public io::net::udp::Receiver::Handler, public Parser:
   // io
   std::unique_ptr<io::net::udp::Receiver> receiver_;
   Reader reader_;
-  Buffer buffer_;
+  tools::Buffer buffer_;
   // status
   std::chrono::nanoseconds last_update_time_ = {};
   ConnectionStatus connection_status_ = {};