+++ src/roq/udp_subscriber/parser.hpp
@@ -7,8 +7,9 @@
 
 #include "roq/core/udp/frame.hpp"
 
-#include "roq/udp_subscriber/header.hpp"
 #include "roq/udp_subscriber/shared.hpp"
+
+#include "roq/udp_subscriber/tools/header.hpp"
 
 namespace roq {
 namespace udp_subscriber {
@@ -16,22 +17,22 @@ namespace udp_subscriber {
 struct Parser {
   struct Heartbeat final {};
   struct Handler {
-    virtual void operator()(Trace<Heartbeat> const &, Header const &) = 0;
-    virtual void operator()(Trace<GatewaySettings> const &, Header const &) = 0;
-    virtual void operator()(Trace<StreamStatus> const &, Header const &) = 0;
-    virtual void operator()(Trace<ExternalLatency> const &, Header const &) = 0;
-    virtual void operator()(Trace<GatewayStatus> const &, Header const &) = 0;
-    virtual void operator()(Trace<ReferenceData> const &, Header const &) = 0;
-    virtual void operator()(Trace<MarketStatus> const &, Header const &) = 0;
-    virtual void operator()(Trace<TopOfBook> const &, Header const &) = 0;
-    virtual void operator()(Trace<MarketByPriceUpdate> const &, Header const &) = 0;
-    virtual void operator()(Trace<TradeSummary> const &, Header const &) = 0;
-    virtual void operator()(Trace<StatisticsUpdate> const &, Header const &) = 0;
-    virtual void operator()(Trace<CustomMetricsUpdate> const &, Header const &) = 0;
+    virtual void operator()(Trace<Heartbeat> const &, tools::Header const &) = 0;
+    virtual void operator()(Trace<GatewaySettings> const &, tools::Header const &) = 0;
+    virtual void operator()(Trace<StreamStatus> const &, tools::Header const &) = 0;
+    virtual void operator()(Trace<ExternalLatency> const &, tools::Header const &) = 0;
+    virtual void operator()(Trace<GatewayStatus> const &, tools::Header const &) = 0;
+    virtual void operator()(Trace<ReferenceData> const &, tools::Header const &) = 0;
+    virtual void operator()(Trace<MarketStatus> const &, tools::Header const &) = 0;
+    virtual void operator()(Trace<TopOfBook> const &, tools::Header const &) = 0;
+    virtual void operator()(Trace<MarketByPriceUpdate> const &, tools::Header const &) = 0;
+    virtual void operator()(Trace<TradeSummary> const &, tools::Header const &) = 0;
+    virtual void operator()(Trace<StatisticsUpdate> const &, tools::Header const &) = 0;
+    virtual void operator()(Trace<CustomMetricsUpdate> const &, tools::Header const &) = 0;
   };
 
   static size_t dispatch(
-      Handler &, Header const &, std::span<std::byte const> const &payload, TraceInfo const &, Shared &);
+      Handler &, tools::Header const &, std::span<std::byte const> const &payload, TraceInfo const &, Shared &);
 };
 
 }  // namespace udp_subscriber