+++ test/dummy.cpp
@@ -1,7 +1,0 @@
-/* Copyright (c) 2017-2022, Hans Erik Thrane */
-
-#include <catch2/catch_all.hpp>
-
-TEST_CASE("simple", "[dummy]") {
-  CHECK(true);
-}